Initial commit of V2.0 of the Moonlight program ===============================================
This is version 2.0 of the Austin et al. moonlight program.
The original code was written for an early version of
Fortran.  In version 2.0 it has been updated to Fortran 90.
The code has been tested using the GNU gfortran compiler on
a linux computer but should compile and run successfully
with other modern fortran compilers on other operating
systems.

Reference:
  Austin, R.H., Phillips, B.F. and Webb, D.J. (1976)  A method
          for calculation moonlight illuminance at the Earth's
          surface.  J. Appl. Ecol., 13(3), 741-748.